[Text Directionality] Fix directionality of non-HTML elements https://bugs.webkit.org/show_bug.cgi?id=276871 rdar://132210868
Reviewed by Ryosuke Niwa.

Move the text directionality code from HTMLElement to Element.

* LayoutTests/imported/w3c/web-platform-tests/html/dom/elements/global-attributes/dir-assorted.window-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/html/dom/elements/global-attributes/dir-auto-dynamic-changes.window-expected.txt:
* Source/WebCore/css/SelectorCheckerTestFunctions.h:
(WebCore::matchesDirPseudoClass):
* Source/WebCore/dom/Element.cpp:
(WebCore::Element::attributeChanged):
(WebCore::parseTextDirection):
(WebCore::isValidDirValue):
(WebCore::elementAffectsDirectionality):
(WebCore::setHasDirAutoFlagRecursively):
(WebCore::Element::hasDirectionAuto const):
(WebCore::Element::directionalityIfDirIsAuto const):
(WebCore::Element::dirAttributeChanged):
(WebCore::Element::updateEffectiveDirectionality):
(WebCore::Element::adjustDirectionalityIfNeededAfterChildrenChanged):
(WebCore::Element::updateTextDirectionalityAfterInputTypeChange):
(WebCore::Element::updateEffectiveDirectionalityOfDirAuto):
(WebCore::Element::computeDirectionalityFromText const):
(WebCore::Element::adjustDirectionalityIfNeededAfterChildAttributeChanged):
(WebCore::Element::insertedIntoAncestor):
(WebCore::Element::removedFromAncestor):
(WebCore::Element::childrenChanged):
* Source/WebCore/dom/Element.h:
* Source/WebCore/html/HTMLElement.cpp:
(WebCore::HTMLElement::attributeChanged):
(WebCore::HTMLElement::insertedIntoAncestor):
(WebCore::HTMLElement::removedFromAncestor):
(WebCore::parseTextDirection): Deleted.
(WebCore::isValidDirValue): Deleted.
(WebCore::elementAffectsDirectionality): Deleted.
(WebCore::setHasDirAutoFlagRecursively): Deleted.
(WebCore::HTMLElement::childrenChanged): Deleted.
(WebCore::HTMLElement::hasDirectionAuto const): Deleted.
(WebCore::HTMLElement::directionalityIfDirIsAuto const): Deleted.
(WebCore::HTMLElement::computeDirectionalityFromText const): Deleted.
(WebCore::HTMLElement::dirAttributeChanged): Deleted.
(WebCore::HTMLElement::updateEffectiveDirectionality): Deleted.
(WebCore::HTMLElement::adjustDirectionalityIfNeededAfterChildAttributeChanged): Deleted.
(WebCore::HTMLElement::updateEffectiveDirectionalityOfDirAuto): Deleted.
(WebCore::HTMLElement::updateTextDirectionalityAfterInputTypeChange): Deleted.
(WebCore::HTMLElement::adjustDirectionalityIfNeededAfterChildrenChanged): Deleted.
* Source/WebCore/html/HTMLElement.h:

Canonical link: https://commits.webkit.org/281998@main
